Move Account and ConstantsWrapper inside auth
